+++ packages/core/src/server/__tests__/funcs/business-500.api.ts
@@ -1,7 +1,7 @@
 import { Func } from '@faasjs/core'
 import { HttpError } from '@faasjs/core'
 
-export const func = new Func({
+export default new Func({
   async handler() {
     throw new HttpError({
       statusCode: 500,