+++ packages/core/src/server/__tests__/funcs/path-alias.api.ts
@@ -6,7 +6,7 @@ import { fromRelative } from './shared/relative'
 
 const http = new Http({ config: { cookie: { session: { secret: 'test-secret' } } } })
 
-export const func = new Func({
+export default new Func({
   plugins: [http],
   async handler() {
     return {