+++ packages/core/src/server/__tests__/funcs/raw.api.ts
@@ -2,7 +2,7 @@ import type { ServerResponse } from 'node:http'
 
 import { Func } from '@faasjs/core'
 
-export const func = new Func({
+export default new Func({
   async handler({ event }) {
     const response = event.raw.response as ServerResponse
     response.statusCode = 200