+++ packages/core/src/server/__tests__/funcs/runtime.api.ts
@@ -4,16 +4,6 @@ import { detectNodeRuntime } from '@faasjs/node-utils'
 
 const http = new Http({ config: { cookie: { session: { secret: 'test-secret' } } } })
 
-export const func = new Func({
-  plugins: [http],
-  async handler() {
-    return {
-      exportOrder: 'func',
-      runtime: detectNodeRuntime(),
-    }
-  },
-})
-
 export default new Func({
   plugins: [http],
   async handler() {