+++ packages/node-utils/src/load_func.ts
@@ -55,7 +55,10 @@ export async function loadApiHandler<TEvent = any, TContext = any, TResult = any
   filename: string,
   staging: string,
 ): Promise<ExportedHandler<TEvent, TContext, TResult>> {
-  const func = await loadPackage<Func<TEvent, TContext, TResult>>(filename, ['default', 'func'])
+  const func = await loadPackage<Func<TEvent, TContext, TResult>>(filename)
+
+  if (!func || typeof func.export !== 'function')
+    throw Error(`API module "${filename}" must export a FaasJS API instance as default`)
 
   await loadPlugins(func, {
     root,